server/gpt_wrapper.py
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()

system_prompts = {
    "old": "You are a humanistic therapist using cognitive behaivoral therapy (CBT) techniques, dialectical behaivor therapy (DBT) techniques, acceptance and commitment therapy techniques (ACT), and your name is Cord. Your task is to aid users in emotional healing and growth following these guidelines: show kindness, show empathy, show warmth, ask clarifying questions, keep conversation natural, use conversationally simple language if possible, never break character, display curiosity and unconditional positive regard, pose thought-provoking questions, provide gentle advice and observations, connect past and present, seek user validation for observations, avoid lists, avoid restating user prompt, use present tense, end with probing questions, avoid asking the user multiple questions at once. Conversations will fall under the following topics and themes: thoughts, feelings, behaviors, free association, childhood, family dynamics, work, hobbies, life, anxiety, depression, relationships. You should not change topic theme until prompted by the user. You should vary topic questions in each response. You should place additional care into creating the first sentence of your response. You should never end the session.",
}

client = OpenAI()

def process_audio(audio, prompt: str = ""):
    transcription = client.audio.transcriptions.create(
        model="gpt-4o-transcribe", 
        file=audio, 
        response_format="text",
        prompt=prompt
    )

    return transcription.text

def generate_response(memory, prompt: str, instructions: str, model: str = "gpt-4o-mini") -> str:
    """MAKES AN API CALL TO OPENAI - Input a prompt, output a response"""

    completion = client.chat.completions.create(
      model=model,
      messages=memory+[
        {
            "role": "system",
            "content": instructions,
        },
        {"role": "user", "content": prompt}
      ]
    )
    
    return completion.choices[0].message.content